+++ packages/capacity-manager/.env.dist
@@ -1,0 +1,44 @@
+# NOTE: This won't work without an AWS deployment at the moment. There is no
+# need to manage capacity when running locally.
+
+# Base Application Configuration
+POLL_INTERVAL_MS=1000
+API_ENDPOINT=http://localhost:5000
+AWS_REGION=ap-southeast-2
+API_USERNAME=worker@email.com
+API_PASSWORD=password
+VPC_ID=vpc-xxxxxxxxx
+
+# Job Type Configuration - Replace with your actual values
+# You'll need to create separate configurations for each JobType enum value
+# Example for a hypothetical "DATA_PROCESSING" job type:
+
+TEST_TASK_DEF=arn:aws:ecs:us-east-1:123456789012:task-definition/data-processing:1
+TEST_CLUSTER=arn:aws:ecs:us-east-1:123456789012:cluster/your-cluster
+TEST_MIN_CAPACITY=0
+TEST_MAX_CAPACITY=10
+TEST_COOLDOWN=300
+TEST_SENSITIVITY=1.5
+TEST_FACTOR=2
+TEST_SECURITY_GROUP=sg-xxxxxxxxx
+
+SUITABILITY_ASSESSMENT_TASK_DEF=arn:aws:ecs:us-east-1:123456789012:task-definition/data-processing:1
+SUITABILITY_ASSESSMENT_CLUSTER=arn:aws:ecs:us-east-1:123456789012:cluster/your-cluster
+SUITABILITY_ASSESSMENT_MIN_CAPACITY=0
+SUITABILITY_ASSESSMENT_MAX_CAPACITY=10
+SUITABILITY_ASSESSMENT_COOLDOWN=300
+SUITABILITY_ASSESSMENT_SENSITIVITY=1.5
+SUITABILITY_ASSESSMENT_FACTOR=2
+SUITABILITY_ASSESSMENT_SECURITY_GROUP=sg-xxxxxxxxx
+
+REGIONAL_ASSESSMENT_TASK_DEF=arn:aws:ecs:us-east-1:123456789012:task-definition/data-processing:1
+REGIONAL_ASSESSMENT_CLUSTER=arn:aws:ecs:us-east-1:123456789012:cluster/your-cluster
+REGIONAL_ASSESSMENT_MIN_CAPACITY=0
+REGIONAL_ASSESSMENT_MAX_CAPACITY=10
+REGIONAL_ASSESSMENT_COOLDOWN=300
+REGIONAL_ASSESSMENT_SENSITIVITY=1.5
+REGIONAL_ASSESSMENT_FACTOR=2
+REGIONAL_ASSESSMENT_SECURITY_GROUP=sg-xxxxxxxxx
+
+# Add more job types as needed based on your JobType enum values
+# Pattern: {JOB_TYPE_NAME}_{PROPERTY}=value
