+++ packages/capacity-manager/src/config.ts
