+++ packages/db/README.md
@@ -697,8 +697,8 @@ Your web-api will now target your DB.
 
 ### Customization
 
-- Modify `src/infra/components/` files to adjust individual service configurations.
-- Update `src/infra/infra.ts` to change overall stack structure.
+- Modify `src/components/` files to adjust individual service configurations.
+- Update `src/infra.ts` to change overall stack structure.
 - Adjust auto-scaling, instance types, and other parameters in the configuration JSON files.
 
 ## Configuring CDK
@@ -1048,14 +1048,14 @@ Example job type configuration:
 const jobTypeSchemas = {
   SUITABILITY_ASSESSMENT: {
     input: z.object({
-      fieldsHere: z.string(),
+      fieldsHere: z.string()
     }),
     result: z
       .object({
-        fieldsHere: z.string(),
+        fieldsHere: z.string()
       })
-      .optional(),
-  },
+      .optional()
+  }
 };
 ```
