+++ packages/example-worker/.env.dist
@@ -1,0 +1,25 @@
+# ReefGuide Test Worker Configuration
+# This is a sample worker implementation for development and testing
+
+# API Connection Settings
+API_ENDPOINT=http://localhost:5000
+WORKER_USERNAME=worker@email.com
+WORKER_PASSWORD=password
+
+# Job Processing Configuration
+# Comma-separated list of job types this worker can handle
+JOB_TYPES=TEST,SUITABILITY_ASSESSMENT,REGIONAL_ASSESSMENT
+
+# Maximum number of jobs to process simultaneously
+MAX_CONCURRENT_JOBS=3
+
+# How often to poll for new jobs (in milliseconds, minimum 1000)
+POLL_INTERVAL_MS=2000
+
+# Worker Behavior Settings
+# How long to wait before shutting down when idle (in milliseconds)
+# Default: 120000 (2 minutes)
+IDLE_TIMEOUT_MS=120000
+
+# Port for health check HTTP server
+PORT=3000
