+++ packages/example-worker/.eslintignore
@@ -1,0 +1,4 @@
+build
+scripts
+dist
+node_modules
