+++ packages/example-worker/.npmignore
@@ -1,0 +1,2 @@
+*.ts
+!*.d.ts
