+++ packages/example-worker/package.json
@@ -1,0 +1,35 @@
+{
+    "name": "@reefguide/example-worker",
+    "version": "0.1",
+    "description": "Example worker implementation for ReefGuide job system.",
+    "main": "src/index.ts",
+    "scripts": {
+        "start": "env-cmd ts-node --files src/index.ts",
+        "watch": "tsc -w",
+        "dev": "env-cmd ts-node-dev --files src/index.ts",
+        "build": "tsc",
+        "lint": "gts lint",
+        "type-check": "tsc --noEmit",
+        "format:write": "prettier --write \"{src,test}/**/*.{ts,js,json,html,scss,md}\" \"!**/node_modules/**\" \"!**/dist/**\"",
+        "format:check": "prettier --check \"{src,test}/**/*.{ts,js,json,html,scss,md}\" \"!**/node_modules/**\" \"!**/dist/**\"",
+        "clean": "gts clean",
+        "fix": "gts fix"
+    },
+    "keywords": [],
+    "author": "AIMS",
+    "license": "MIT",
+    "dependencies": {
+        "@reefguide/db": "workspace:*",
+        "@reefguide/types": "workspace:*",
+        "axios": "^1.7.7",
+        "env-cmd": "^10.1.0",
+        "express": "^4.21.1",
+        "jwt-decode": "^4.0.0",
+        "zod": "^3.23.8"
+    },
+    "devDependencies": {
+        "@types/express": "^4.17.21",
+        "ts-node": "^10.9.2",
+        "ts-node-dev": "^2.0.0"
+    }
+}
