+++ packages/example-worker/src/authClient.ts
