+++ packages/example-worker/src/config.ts
