+++ packages/example-worker/src/ecs.ts
