+++ packages/web-api/README.md
@@ -697,8 +697,8 @@ Your web-api will now target your DB.
 
 ### Customization
 
-- Modify `src/infra/components/` files to adjust individual service configurations.
-- Update `src/infra/infra.ts` to change overall stack structure.
+- Modify `src/components/` files to adjust individual service configurations.
+- Update `src/infra.ts` to change overall stack structure.
 - Adjust auto-scaling, instance types, and other parameters in the configuration JSON files.
 
 ## Configuring CDK
